refactor: auth内を整理（機能に影響なし）
